devops_bot/bot/requirements.txt
python-telegram-bot==13.7
paramiko
psycopg2-binary
python-dotenv
docker